[NET-512] Rename manager plurals
