+++ cmdb/__update__.py
@@ -73,7 +73,7 @@ class UpdateRoutine:
         if not self.__check_database():
             self.status = UpdateRoutine.UpateStatus.ERROR
             raise RuntimeError(
-                'The database manager could not be initialized. Perhaps the database cannot be reached, \
+                'The database managers could not be initialized. Perhaps the database cannot be reached, \
                 or the database was already initialized.'
             )
