+++ cmdb/framework/cmdb_base.py
@@ -25,17 +25,17 @@ LOGGER = logging.getLogger(__name__)
 
 class CmdbManagerBase(ABC):
     """Represents the base class for cmdb managers. A respective implementation is always adapted to the
-       respective database manager :class:`cmdb.data_storage.DatabaseManager`.
+       respective database managers :class:`cmdb.data_storage.DatabaseManager`.
        But should always use at least the super functions listed here.
     """
 
     def __init__(self, database_manager: DatabaseManagerMongo):
         """Example:
-            Depending on the condition or whether a fork process is used, the database manager can also be seen
-            directly in the declaration of the object manager
+            Depending on the condition or whether a fork process is used, the database managers can also be seen
+            directly in the declaration of the object managers
 
         Args:
-            database_manager (DatabaseManager): initialisation of an database manager
+            database_manager (DatabaseManager): initialisation of an database managers
 
         """
         self.dbm: DatabaseManagerMongo = database_manager
