+++ cmdb/framework/managers/error/__init__.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-from cmdb.framework.manager.error.manager_errors import ManagerGetError, ManagerInsertError, ManagerUpdateError, \
+from cmdb.framework.managers.error.manager_errors import ManagerGetError, ManagerInsertError, ManagerUpdateError, \
     ManagerDeleteError
 
 __all__ = [
