+++ cmdb/framework/managers/results.py
