+++ cmdb/interface/net_app/__init__.py
@@ -35,7 +35,7 @@ system_config_reader = SystemConfigReader()
 
 
 def create_app(event_queue):
-    # Create manager
+    # Create managers
     from cmdb.data_storage.database_manager import DatabaseManagerMongo
     app_database = DatabaseManagerMongo(
         **system_config_reader.get_all_values_from_section('Database')
