+++ cmdb/interface/rest_api/framework_routes/category_routes.py
@@ -19,10 +19,10 @@ import logging
 from flask import abort, current_app, request
 
 from cmdb.framework.models.category import CategoryModel, CategoryTree
-from cmdb.framework.manager import ManagerGetError, ManagerInsertError, ManagerDeleteError, ManagerUpdateError
-from cmdb.framework.manager.category_manager import CategoryManager
-from cmdb.framework.manager.error.framework_errors import FrameworkIterationError
-from cmdb.framework.manager.results import IterationResult
+from cmdb.framework.managers import ManagerGetError, ManagerInsertError, ManagerDeleteError, ManagerUpdateError
+from cmdb.framework.managers.category_manager import CategoryManager
+from cmdb.framework.managers.error.framework_errors import FrameworkIterationError
+from cmdb.framework.managers.results import IterationResult
 from cmdb.framework.utils import PublicID
 from cmdb.interface.api_parameters import CollectionParameters
 from cmdb.interface.response import GetSingleResponse, GetMultiResponse, InsertSingleResponse, DeleteSingleResponse, \
